requirements.txt
aioitertools==0.12.0
annotated-types==0.7.0
anyio==4.9.0
ase==3.26.0
attrs==25.3.0
authlib==1.6.0
bcrypt==4.3.0
bibtexparser==1.4.3
blinker==1.9.0
boto3==1.38.38
botocore==1.38.38
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dnspython==2.7.0
duckdb==1.3.1
emmet-core==0.84.8
exceptiongroup==1.3.0
fastmcp==2.8.1
flask==3.1.1
fonttools==4.58.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
latexcodec==3.0.1
maggma==0.71.5
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mcp==1.9.4
mdurl==0.1.2
mongomock==4.3.0
monty==2025.3.3
mp-api==0.45.7
mpmath==1.3.0
msgpack==1.1.1
narwhals==1.43.0
networkx==3.5
numpy==2.3.0
nvidia-nccl-cu12==2.27.7
openapi-pydantic==0.5.1
orjson==3.10.18
packaging==25.0
palettable==3.3.3
pandas==2.3.0
paramiko==3.5.1
pillow==11.2.1
plotly==6.1.2
pybtex==0.24.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydash==8.0.5
pygments==2.19.1
pymatgen==2025.6.14
pymongo==4.10.1
pynacl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.25.1
ruamel-yaml==0.18.14
ruamel-yaml-clib==0.2.12
s3transfer==0.13.0
scikit-learn==1.7.1
scipy==1.15.3
sentinels==1.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
some-package==0.1
spglib==2.6.0
sse-starlette==2.3.6
sshtunnel==0.4.0
starlette==0.47.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
uncertainties==3.2.3
urllib3==2.4.0
uvicorn==0.34.3
vtk==9.4.2
werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.3
